[add] call expression hint for callequation
